Fix frozen string error. Aligned more with GitHub rendering (#8)
* Dupe the content before modifying.

This fixes an issue where we get: Error:  can't modify frozen String: ""

* Return early if the doc content is empty

Empty strings are frozen by default

* Ignore the `.DS_Store` folder

* Handle empty titles

0cba248e5cbb57d649e2dca893af5c7c8900400e introduced a bug when rendering titles.
First, it doesn't handle empty titles well. Second, the regex was too aggressive. It needs to stop at the newline.

This fixes that.

* The blockquote character should be at beginning of string

* Fix newlines in the block quote

* Handle indentation of block quotes

* Handle markdown after admonitions

The admonitions regex was capturing newlines after it. We need to give that one back.

* Don't add newlines

Follow the GitHub .md behavior, not the behavior in GitHub issues/pr.